feat: add timestamp

+++ internal/core/unittest/test_cached_search_iterator.cpp
@@ -251,7 +251,7 @@ class CachedSearchIteratorTest
 
         size_t offset = 0;
         for (size_t i = 0; i < num_chunks_; ++i) {
-            const size_t rows = std::min(nb_ - offset, kSizePerChunk);
+            const size_t rows = std::min(static_cast<size_t>(nb_ - offset), kSizePerChunk);
             const size_t buf_size = rows * dim_ * sizeof(float);
             auto& chunk_data = column_data_[i];
             chunk_data.resize(buf_size);
